update 0224 Nnovels mada

+++ src/views/admin/AdminReports.vue
@@ -1,24 +1,25 @@
 <template lang="pug">
-b-container#adminissues.my-5
+b-container#adminreports.my-5
   b-row.mb-5
     b-col.h1.rounded.text-white.text-center.py-3.aino-bg-primary.aino-rounded(cols='12') 問題回報管理
   b-table(striped hover :items='reports' :fields='fields' ref='table')
     template(#cell(processState)='data')
       | {{ data.item.processed ? '已處理完成' : '尚未處理' }}
     template(#cell(detail)='data')
-      b-btn.aino-btn-third.aino-bg-third(@click='adminReportsDetail(data.index)') 詳細狀況
+      b-btn.aino-btn-third(@click='adminReportsDetail(data.index)') 詳細狀況
     template(#cell(delete)='data')
-      b-btn.aino-btn-third.bg-danger(@click='adminDeleteReports(data.index)') 刪除
+      b-btn.aino-btn-danger(@click='adminDeleteReports(data.index)') 刪除
 
   b-modal#modal-adminReportsDetail(
     size="lg"
     centered
     title='詳細狀況'
     ok-title='處理完成'
-    cancel-title='取消編輯'
+    cancel-title='尚未處理'
     ok-variant='success'
     cancel-variant='danger'
-    @ok=''
+    @ok='reportsState(true)'
+    @cancel='reportsState(false)'
   )
     b-row.border-bottom
       b-col.d-flex.h5(cols='12')
@@ -50,6 +51,8 @@ b-container#adminissues.my-5
     cancel-variant='success'
     @ok='deleteReports'
   )
+    .h2.deleteSignal.mx-auto.my-3
+      div.deleteSignalText !
     .h2.text-center 是否刪除此回報?
 </template>
 
@@ -80,7 +83,38 @@ export default {
       this.form = { ...this.reports[index], index }
       this.$bvModal.show('modal-adminDeleteReports')
     },
+    async reportsState (value) {
+      try {
+        const { data } = await this.api.patch('/reports/reportstate/' + this.form._id, { processed: value }, {
+          headers: {
+            authorization: 'Bearer ' + this.user.token
+          }
+        })
+        console.log(data.result)
+        this.$swal({
+          icon: 'success',
+          title: '成功',
+          text: '此回報問題已完成處理'
+        })
+        this.reports[this.form.index] = { ...this.form, processed: value }
+        this.$refs.table.refresh()
+      } catch (error) {
+        this.$swal({
+          icon: 'error',
+          title: '錯誤',
+          text: '處理狀態變更失敗'
+        })
+      }
+    },
     async deleteReports () {
+      if (!this.form.processed) {
+        this.$swal({
+          icon: 'error',
+          title: '錯誤',
+          text: '此回報尚未處理完成'
+        })
+        return
+      }
       try {
         const { data } = await this.api.delete('/reports/reportsdelete/' + this.form._id, {
           headers: {
@@ -124,12 +158,6 @@ export default {
 </script>
 
 <style lang="scss">
-@import '../../../scss/primary.scss' ;
+@import '../../../scss/aino-style.scss' ;
 
-.reportTextBox {
-  width: 100%;
-  border: 1px solid $font-second;
-  border-radius: 10px;
-  color: $font-primary;
-}
 </style>
